backend/requirement.txt
llama-index==0.8.54
langchain==0.0.352
openai==0.28.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
numpy==1.26.2
pydantic>=2.5.2
PyMUPDF==1.23.7